+++ app/components/searchworks4/search_result/bookplate_fund_component.html.erb
