+++ app/components/searchworks4/search_result/collection_info_component.html.erb
